+++ packages/server/src/lib/filename-to-id.ts
@@ -1,3 +1,6 @@
+import path from "path";
+import { STORAGE_FILE_EXTENSION } from "./id-to-filename";
+
 export function filenameToId(filename: string) {
-  return filename.split(".")[0];
+  return path.basename(filename, STORAGE_FILE_EXTENSION);
 }