+++ packages/server/src/lib/id-to-filename.ts
@@ -1,5 +1,5 @@
-export const STORAGE_FILE_EXTENSION = "haiku";
+export const STORAGE_FILE_EXTENSION = ".haiku";
 
 export function idToFilename(id: string) {
-  return `${id}.${STORAGE_FILE_EXTENSION}`;
+  return `${id}${STORAGE_FILE_EXTENSION}`;
 }