+++ packages/server/src/routes/get-incoming-links.ts
@@ -32,7 +32,7 @@ export const handleGetIncomingLinks = createHandler<GetIncomingLinksOuput, GetIn
 async function getIncomingLinks(id: string): Promise<IncomingLink[]> {
   const config = await getRepoMetadata();
 
-  const { error, stdout, stderr } = await runShell(`rg "\(${id}\)" --count-matches`, { cwd: config.repoDir });
+  const { error, stdout, stderr } = await runShell(`rg "\(${id}\)" --count-matches ./`, { cwd: config.repoDir });
 
   if (error) {
     if (error.code === 1) {